pyproject.toml
[tool.poetry]
name = "causy"
version = "0.1.0"
description = "Causal discovery made easy."
authors = ["Sofia Faltenbacher <faltenbacher.sofia@gmail.com>", "Lilith Wittmann <mail@lilithwittmann.de>"]
readme = "README.md"
license = "MIT"
include = ["pyproject.toml", "README.md", "LICENSE", "causy", "causy/static/*", "causy/static/assets/*"]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
scipy = "^1.13"
typer = {extras = ["all"], version = "^0.12.3"}
torch = "^2.1.0"
fastapi = "^0.110.2"
uvicorn = "^0.27.0"
srsly = "^2.4.8"
pydantic-yaml = "^1.2.1"
click = "^8.1.7"
questionary = "^2.0.1"
jinja2 = "^3.1.4"
markdown = "^3.6"
httpx = "^0.27.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.5.1"
pyinstrument = "^4.5.3"
black = "^23.9.1"
numpy = "^1.26.0"
pdoc = "^14.1.0"
pre-commit = "^3.4.0"
coverage = "^7.3.2"


[tool.poetry-version-plugin]
source = "git-tag"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
causy = "causy.cli:app"
